Another fix for compiling in 64 bit linux.
git-svn-id: file:///dev/shm/opensim/Trunk@2694 fc29be9c-2cb0-420d-b520-73dbd18add6e

+++ OpenSim/Simulation/SIMM/SimmPathMatrix.cpp
@@ -183,7 +183,7 @@ void SimmPathMatrix::invalidate()
  */
 int SimmPathMatrix::hash(const AbstractBody* aFromBody, const AbstractBody* aToBody) const
 {
-	int hash_value = ((int)aFromBody / cHash1 + (int)aToBody) / cHash2 % _size;
+	int hash_value = ((int)(intptr_t)aFromBody / cHash1 + (int)(intptr_t)aToBody) / cHash2 % _size;
 
 	SimmPath* hashEntry;
 	for (int i = 0; i < _size; i++)
